+++ src/routes/+layout.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { faGamepad } from "@fortawesome/free-solid-svg-icons";
+  import { faGear } from "@fortawesome/free-solid-svg-icons";
   import "bootstrap/dist/css/bootstrap.min.css";
   import Fa from "svelte-fa";
   import {
@@ -15,7 +15,15 @@
     NavbarToggler,
   } from "sveltestrap";
 
-  import { completeFilter, ignoredFilter, incompleteFilter, nameFilter, oneRegionFilter } from "../store.js";
+  import SettingsModal from "../components/SettingsModal.svelte";
+  import {
+    completeFilter,
+    ignoredFilter,
+    incompleteFilter,
+    isSettingsModalOpen,
+    nameFilter,
+    oneRegionFilter,
+  } from "../store.js";
 
   let navbarIsOpen = false;
 
@@ -41,7 +49,7 @@
       <ButtonToolbar class="d-flex">
         <ButtonGroup class="ms-3">
           <Button color="primary" bind:active={$oneRegionFilter} on:click={() => oneRegionFilter.update((b) => !b)}>
-            {#if $oneRegionFilter}Show{:else}Hide{/if} 1G1R
+            {#if $oneRegionFilter}Show All{:else}Show 1G1R only{/if}
           </Button>
         </ButtonGroup>
         <ButtonGroup class="ms-3">
@@ -58,11 +66,22 @@
         <InputGroup class="ms-3">
           <Input placeholder="Game Name" bind:value={$nameFilter} />
         </InputGroup>
+        <ButtonGroup class="ms-3">
+          <Button
+            color="dark"
+            bind:active={$isSettingsModalOpen}
+            on:click={() => isSettingsModalOpen.update((b) => !b)}
+          >
+            <Fa icon={faGear} />
+          </Button>
+        </ButtonGroup>
       </ButtonToolbar>
-    </Collapse> />
+    </Collapse>
   </Navbar>
 
   <Container fluid class="flex-fill">
     <slot />
   </Container>
+
+  <SettingsModal toggle={() => isSettingsModalOpen.update((b) => !b)} />
 </div>